Revert "New format for url and new router schema without plist"
