+++ Example/BXRouter/BXViewController.m
@@ -26,25 +26,23 @@
         [array addObject:[[BXRouterMapItem alloc] initWithObject:item]];
     }
     [[BXRouterManager shareVCManager] registerRouterMapList:array];
-    [[BXRouterManager shareVCManager] registerClassPrefix:@"BX"];
 }
 
 - (IBAction)jumpByStoryboard:(id)sender
 {
-    BXRouterUrl *url = [[BXRouterUrl alloc] initWithUrl:@"bxapp://name=Storyboard&category=storyboard:Main&transform=push/paramA=xxx&paramB=xxx"];
+    BXRouterUrl *url = [[BXRouterUrl alloc] initWithUrl:@"bxapp://storyboard_controller/?"];
     [[BXRouterManager shareVCManager] openUrl:url delegate:self];
 }
 
 - (IBAction)jumpByNib:(id)sender
 {
-    BXRouterUrl *url = [[BXRouterUrl alloc] initWithUrl:@"bxapp://name=NibView&transform=present/paramA=xxx&paramB=xxx"];
+    BXRouterUrl *url = [[BXRouterUrl alloc] initWithUrl:@"bxapp://nib_controller/?"];
     [[BXRouterManager shareVCManager] openUrl:url delegate:self];
 }
 
 - (IBAction)jumpByCode:(id)sender
 {
-    [[BXRouterManager shareVCManager] resetClassPrefix];
-    BXRouterUrl *url = [[BXRouterUrl alloc] initWithUrl:@"bxapp://name=BXCodeViewController/?"];
+    BXRouterUrl *url = [[BXRouterUrl alloc] initWithUrl:@"bxapp://code_controller/?"];
     [[BXRouterManager shareVCManager] openUrl:url delegate:self];
 }
