+++ Pod/Classes/BXRouterManager.m
@@ -7,15 +7,12 @@
 //
 
 #import "BXRouterManager.h"
-#import "BXRouterConfig.h"
 #import "BXRouterMapItem.h"
 #import <objc/runtime.h>
 
 @interface BXRouterManager ()
 
 @property (nonatomic, strong) NSArray *vcMap;
-
-@property (nonatomic, strong, readwrite) NSString *classPrefix;
 
 @end
 
@@ -46,26 +43,13 @@
     return [BXRouterManager shareVCManager];
 }
 
-- (BOOL)registerClassPrefix:(NSString *)prefix
-{
-    self.classPrefix = prefix;
-
-    return YES;
-}
-
-- (BOOL)resetClassPrefix {
-    self.classPrefix = @"";
-    
-    return YES;
-}
-
 - (BOOL)registerRouterMapList:(NSArray *)routerList
 {
     __block BOOL verify = YES;
     
     [routerList enumerateObjectsUsingBlock:^(id obj, NSUInteger idx, BOOL *stop) {
         if (![obj isMemberOfClass:[BXRouterMapItem class]]) {
-            *stop  = YES;
+            *stop = YES;
             verify = NO;
         }
     }];
@@ -79,8 +63,38 @@
     return verify;
 }
 
-- (UIViewController<BXRouterProtocol> *)configureControllerByPlist:(BXRouterUrl *)url {
-    BXRouterMapItem *mapItem = [self mapItemByAlias:url.classAlias];
+- (BXRouterMapItem *)mapItemByAlias:(NSString *)alias
+{
+    for (BXRouterMapItem *mapItem in self.vcMap) {
+        if ([mapItem.alias compare:alias] == NSOrderedSame) {
+            return mapItem;
+        }
+    }
+    return nil;
+}
+
+- (BXTransformType)getTransformTypeByAlias:(NSString *)alias
+{
+    // default is push transition.
+    BXTransformType transformType = BXTransformPush;
+    
+    for (BXRouterMapItem *mapItem in _vcMap) {
+        if ([mapItem.alias compare:alias] == NSOrderedSame) {
+            if ([mapItem.transform isEqualToString:@"present"]) {
+                transformType = BXTransformPresent;
+            } else if ([mapItem.transform isEqualToString:@"pop"]) {
+                transformType = BXTransformPop;
+            }
+            break;
+        }
+    }
+    
+    return transformType;
+}
+
+- (UIViewController<BXRouterProtocol> *)getControllerByUrl:(BXRouterUrl *)url
+{
+    BXRouterMapItem *mapItem = [self mapItemByAlias:url.vcAlias];
     
     if ([[NSBundle mainBundle] pathForResource:mapItem.vcClass ofType:@"nib"]) {
         // return view controller from nib
@@ -97,41 +111,8 @@
             }
         }
     }
-    return nil;
-}
-
-- (BXRouterMapItem *)mapItemByAlias:(NSString *)alias
-{
-    for (BXRouterMapItem *mapItem in self.vcMap) {
-        if ([mapItem.alias compare:alias] == NSOrderedSame) {
-            return mapItem;
-        }
-    }
-    return nil;
-}
-
-- (BXTransformType)configureTransformTypeByPlist:(NSString *)alias delegate:(UIViewController *)delegate
-{
-    // default is push transition.
-    BXTransformType transformType = BXTransformPush;
     
-    for (BXRouterMapItem *mapItem in _vcMap) {
-        if ([mapItem.alias compare:alias] == NSOrderedSame) {
-            if ([mapItem.transform isEqualToString:@"present"]) {
-                transformType = BXTransformPresent;
-            } else if ([mapItem.transform isEqualToString:@"pop"]) {
-                transformType = BXTransformPop;
-            }
-            break;
-        }
-    }
-    if (nil == delegate.navigationController) {
-        // if navigationController is nil, must present,
-        // or adding a root view controller as a child of view controller
-        transformType = BXTransformPresent;
-    }
-    
-    return transformType;
+    return nil;
 }
 
 - (UIViewController<BXRouterProtocol> *)openUrl:(BXRouterUrl *)url delegate:(UIViewController *)delegate
@@ -139,20 +120,8 @@
     NSAssert(self.vcMap, @"please call 'registerRouterMapList' method first.");
     
     // get controller by url
-    UIViewController<BXRouterProtocol> *controller = [[BXRouterConfig shareConfig]
-                                                        configureControllerByUrl:url
-                                                            withPrefix:self.classPrefix];
-    // get transform type
-    BXTransformType transform = [[BXRouterConfig shareConfig] configureTransformTypeByUrl:url
-                                                                             withDelegate:delegate];
-
-    // can not get controller by parsing class alias, to search in plist.
-    if (nil == controller) {
-        controller = [self configureControllerByPlist:url];
-        NSAssert(controller, @"can not find the class by alias or plist");
-        
-        transform  = [self configureTransformTypeByPlist:url.classAlias delegate:delegate];
-    }
+    UIViewController<BXRouterProtocol> *controller = [self getControllerByUrl:url];
+    
     // parse router url queryParams
     if ([controller respondsToSelector:@selector(queryParamsToPropertyKeyPaths:)]) {
         [controller queryParamsToPropertyKeyPaths:url];
@@ -164,10 +133,19 @@
         needInNavigationController = [controller needInNavigationController];
     }
     
+    // get transform type
+    BXTransformType transform = BXTransformNone;
+    if (nil == delegate.navigationController) {
+        // if navigationController is nil, must present,
+        // or adding a root view controller as a child of view controller
+        transform = BXTransformPresent;
+    } else {
+        transform = [self getTransformTypeByAlias:url.vcAlias];
+    }
+    
     if (BXTransformPresent == transform) {
         if (needInNavigationController) {
-            UINavigationController *navigation = [[UINavigationController alloc]
-                                                  initWithRootViewController:controller];
+            UINavigationController *navigation = [[UINavigationController alloc] initWithRootViewController:controller];
             [delegate presentViewController:navigation animated:YES completion:^{}];
         } else {
             [delegate presentViewController:controller animated:YES completion:^{}];
@@ -179,7 +157,7 @@
             [delegate.navigationController pushViewController:controller animated:YES];
         }
     } else {
-        BXRouterMapItem *mapItem = [self mapItemByAlias:url.classAlias];
+        BXRouterMapItem *mapItem = [self mapItemByAlias:url.vcAlias];
         __weak NSArray *viewControllers = delegate.navigationController.viewControllers;
         [viewControllers enumerateObjectsUsingBlock:^(id obj, NSUInteger idx, BOOL *stop) {
             if ([obj isMemberOfClass:NSClassFromString(mapItem.vcClass)]) {
@@ -190,6 +168,7 @@
             }
         }];
     }
+    
     return controller;
 }
